@@ -9,12 +9,27 @@
     <version>1.0.0-SNAPSHOT</version>
 
     <properties>
-        <jdk.version>1.8</jdk.version>
-        <kotlin.version>1.3.71</kotlin.version>
+        <!-- language -->
+        <java.version>11</java.version>
+        <kotlin.version>1.4.10</kotlin.version>
+        <kotlin.compiler.jvmTarget>11</kotlin.compiler.jvmTarget>
         <kotlinx.version>1.3.7</kotlinx.version>
+        <maven.compiler.source>11</maven.compiler.source>
+        <maven.compiler.target>11</maven.compiler.target>
 
-        <kotlin-logging.version>1.4.7</kotlin-logging.version>
-        <slf4j.version>1.7.2</slf4j.version>
+        <!-- logging -->
+        <kotlin-logging.version>1.6.24</kotlin-logging.version>
+        <logback.version>1.2.3</logback.version>
+
+        <!-- testing -->
+        <assertj.version>3.6.1</assertj.version>
+        <mokito.version>2.1.0</mokito.version>
+        <testcontainers.version>1.15.3</testcontainers.version>
+        <junit-platform.version>5.7.2</junit-platform.version>
+        <surefire-plugin.version>2.22.2</surefire-plugin.version>
+        <surefire-exclude-test-tags>test-containers</surefire-exclude-test-tags>
+        <ktor.version>1.6.1</ktor.version>
+        <h2.version>1.4.200</h2.version>
 
         <jackson.version>2.11.1</jackson.version>
         <sparkjava.version>2.7.2</sparkjava.version>
