@@ -1,7 +1,7 @@
 package me.roybailey.research.kotlin.basic
 
 import org.assertj.core.api.SoftAssertions
-import org.junit.Test
+import org.junit.jupiter.api.Test
 
 
 class KotlinCollections {
