@@ -1,7 +1,7 @@
 package me.roybailey.research.kotlin.dsl
 
 import org.assertj.core.api.Assertions.assertThat
-import org.junit.Test
+import org.junit.jupiter.api.Test
 
 
 // ------------------------------------------------------------
