@@ -1,7 +1,8 @@
 package me.roybailey.research.kotlin.h2
 
 import mu.KotlinLogging
-import org.junit.Test
+
+import org.junit.jupiter.api.Test
 import java.io.File
 import java.sql.Connection
 import java.sql.DriverManager
